imuxsock: add ability to NOT create/delete sockets during startup and shutdown
closes: http://bugzilla.adiscon.com/show_bug.cgi?id=259

+++ plugins/imuxsock/imuxsock.c
@@ -146,6 +146,7 @@ typedef struct lstn_s {
 	sbool bWritePid;	/* write original PID into tag */
 	sbool bDiscardOwnMsgs;	/* discard messages that originated from ourselves */
 	sbool bUseSysTimeStamp;	/* use timestamp from system (instead of from message) */
+	sbool bUnlink;		/* unlink&re-create socket at start and end of processing */
 } lstn_t;
 static lstn_t listeners[MAXFUNIX];
 
@@ -201,6 +202,7 @@ struct instanceConf_s {
 	int bAnnotate;			/* annotate trusted properties */
 	int bParseTrusted;		/* parse trusted properties */
 	sbool bDiscardOwnMsgs;		/* discard messages that originated from our own pid? */
+	sbool bUnlink;
 	struct instanceConf_s *next;
 };
 
@@ -220,6 +222,7 @@ struct modConfData_s {
 	sbool bUseSysTimeStamp;
 	sbool bDiscardOwnMsgs;
 	sbool configSetViaV2Method;
+	sbool bUnlink;
 };
 static modConfData_t *loadModConf = NULL;/* modConf ptr to use for the current load process */
 static modConfData_t *runModConf = NULL;/* modConf ptr to use for the current load process */
@@ -228,6 +231,7 @@ static modConfData_t *runModConf = NULL;/* modConf ptr to use for the current lo
 static struct cnfparamdescr modpdescr[] = {
 	{ "syssock.use", eCmdHdlrBinary, 0 },
 	{ "syssock.name", eCmdHdlrGetWord, 0 },
+	{ "syssock.unlink", eCmdHdlrBinary, 0 },
 	{ "syssock.ignoretimestamp", eCmdHdlrBinary, 0 },
 	{ "syssock.ignoreownmessages", eCmdHdlrBinary, 0 },
 	{ "syssock.flowcontrol", eCmdHdlrBinary, 0 },
@@ -248,6 +252,7 @@ static struct cnfparamblk modpblk =
 /* input instance parameters */
 static struct cnfparamdescr inppdescr[] = {
 	{ "socket", eCmdHdlrString, CNFPARAM_REQUIRED }, /* legacy: addunixlistensocket */
+	{ "unlink", eCmdHdlrBinary, 0 },
 	{ "createpath", eCmdHdlrBinary, 0 },
 	{ "parsetrusted", eCmdHdlrBinary, 0 },
 	{ "ignoreownmessages", eCmdHdlrBinary, 0 },
@@ -295,6 +300,7 @@ createInstance(instanceConf_t **pinst)
 	inst->bAnnotate = 0;
 	inst->bParseTrusted = 0;
 	inst->bDiscardOwnMsgs = 1;
+	inst->bUnlink = 1;
 	inst->next = NULL;
 
 	/* node created, let's add to config */
@@ -399,6 +405,7 @@ addListner(instanceConf_t *inst)
 		listeners[nfd].bAnnotate = inst->bAnnotate;
 		listeners[nfd].bParseTrusted = inst->bParseTrusted;
 		listeners[nfd].bDiscardOwnMsgs = inst->bDiscardOwnMsgs;
+		listeners[nfd].bUnlink = inst->bUnlink;
 		listeners[nfd].bWritePid = inst->bWritePid;
 		listeners[nfd].bUseSysTimeStamp = inst->bUseSysTimeStamp;
 		CHKiRet(ratelimitNew(&listeners[nfd].dflt_ratelimiter, "imuxsock", NULL));
@@ -449,7 +456,8 @@ createLogSocket(lstn_t *pLstn)
 	struct sockaddr_un sunx;
 	DEFiRet;
 
-	unlink((char*)pLstn->sockName);
+	if(pLstn->bUnlink)
+		unlink((char*)pLstn->sockName);
 	memset(&sunx, 0, sizeof(sunx));
 	sunx.sun_family = AF_UNIX;
 	if(pLstn->bCreatePath) {
@@ -1055,6 +1063,7 @@ activateListeners()
 	listeners[0].bAnnotate = runModConf->bAnnotateSysSock;
 	listeners[0].bParseTrusted = runModConf->bParseTrusted;
 	listeners[0].bDiscardOwnMsgs = runModConf->bDiscardOwnMsgs;
+	listeners[0].bUnlink = runModConf->bUnlink;
 	listeners[0].bUseSysTimeStamp = runModConf->bUseSysTimeStamp;
 	listeners[0].flags = runModConf->bIgnoreTimestamp ? IGNDATE : NOFLAG;
 	listeners[0].flowCtl = runModConf->bUseFlowCtl ? eFLOWCTL_LIGHT_DELAY : eFLOWCTL_NO_DELAY;
@@ -1104,6 +1113,7 @@ CODESTARTbeginCnfLoad
 	pModConf->bAnnotateSysSock = 0;
 	pModConf->bParseTrusted = 0;
 	pModConf->bDiscardOwnMsgs = 1;
+	pModConf->bUnlink = 1;
 	pModConf->ratelimitIntervalSysSock = DFLT_ratelimitInterval;
 	pModConf->ratelimitBurstSysSock = DFLT_ratelimitBurst;
 	pModConf->ratelimitSeveritySysSock = DFLT_ratelimitSeverity;
@@ -1140,6 +1150,8 @@ CODESTARTsetModCnf
 			loadModConf->bIgnoreTimestamp = (int) pvals[i].val.d.n;
 		} else if(!strcmp(modpblk.descr[i].name, "syssock.ignoreownmessages")) {
 			loadModConf->bDiscardOwnMsgs = (int) pvals[i].val.d.n;
+		} else if(!strcmp(modpblk.descr[i].name, "syssock.unlink")) {
+			loadModConf->bUnlink = (int) pvals[i].val.d.n;
 		} else if(!strcmp(modpblk.descr[i].name, "syssock.flowcontrol")) {
 			loadModConf->bUseFlowCtl = (int) pvals[i].val.d.n;
 		} else if(!strcmp(modpblk.descr[i].name, "syssock.usesystimestamp")) {
@@ -1204,6 +1216,8 @@ CODESTARTnewInpInst
 			inst->bParseTrusted = (int) pvals[i].val.d.n;
 		} else if(!strcmp(modpblk.descr[i].name, "ignoreownmessages")) {
 			inst->bDiscardOwnMsgs = (int) pvals[i].val.d.n;
+		} else if(!strcmp(modpblk.descr[i].name, "unlink")) {
+			inst->bUnlink = (int) pvals[i].val.d.n;
 		} else if(!strcmp(modpblk.descr[i].name, "hostname")) {
 			inst->pLogHostName = (uchar*)es_str2cstr(pvals[i].val.d.estr, NULL);
 		} else if(!strcmp(modpblk.descr[i].name, "ignoretimestamp")) {
@@ -1383,8 +1397,10 @@ CODESTARTafterRun
 			    listeners[i].fd <  SD_LISTEN_FDS_START + sd_fds)
 				continue;
 
-			DBGPRINTF("imuxsock: unlinking unix socket file[%d] %s\n", i, listeners[i].sockName);
-			unlink((char*) listeners[i].sockName);
+			if(listeners[i].bUnlink) {
+				DBGPRINTF("imuxsock: unlinking unix socket file[%d] %s\n", i, listeners[i].sockName);
+				unlink((char*) listeners[i].sockName);
+			}
 		}
 
 	discardLogSockets();
@@ -1497,6 +1513,7 @@ CODEmodInit_QueryRegCFSLineHdlr
 	listeners[0].bAnnotate = 0;
 	listeners[0].bParseTrusted = 0;
 	listeners[0].bDiscardOwnMsgs = 1;
+	listeners[0].bUnlink = 1;
 	listeners[0].bCreatePath = 0;
 	listeners[0].bUseSysTimeStamp = 1;
 	if((listeners[0].ht = create_hashtable(100, hash_from_key_fn, key_equals_fn,
